FV-43 enhance fetching image list
Issue
When going to the image overview, the site loading is very slow. This is due to fact, that all images have been loaded from the databse.

Enhancement
Loading all images is not necessary, since only the name and database id is required. As such, a JPA projection "cuts" only the id and filename from the JPA entity in order to speed up loading.